example/mockups_project/requirements.txt
django-avatar>=8.0.1,<8.1
lorem-text>=2.1,<2.2